requirements.txt
asyncssh==2.20.0
flask
flask-cors
flask_socketio
flask-restx
flask-cors
python-dotenv
pydantic~=2.10.6
openai~=1.66.3
tenacity~=9.0.0
pyyaml~=6.0.2
loguru~=0.7.3
tiktoken~=0.9.0
aiofiles~=24.1.0
pydantic_core~=2.27.2
docker~=7.1.0
pytest~=8.3.5
pytest-asyncio~=0.25.3
mcp~=1.4.1
httpx>=0.27.0
boto3~=1.37.16

# numpy
# datasets~=3.2.0
fastapi~=0.115.11
# gymnasium~=1.0.0
# pillow~=10.4.0
# browsergym~=0.13.3
uvicorn~=0.34.0
# unidiff~=0.7.5
# colorama~=0.4.6
# tomli>=2.0.0